Fix errors reported by phpcs
