+++ test/unit/Command/ValidationCommandDummy.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * @package: chapi
+ *
+ * @author:  ppokatilo
+ * @since:   2017-08-84
+ */
+
+namespace unit\Command;
+
+use Chapi\Commands\ValidationCommand;
+
+class ValidationCommandDummy extends ValidationCommand
+{
+    public static $oContainerDummy;
+
+    protected function getContainer()
+    {
+        return self::$oContainerDummy;
+    }
+
+    protected function isAppRunable()
+    {
+        return true;
+    }
+}
